Scheduler for Internal Medicine internship
This simple script tries to find a schedule that best fits everybody's needs for our internship at NCKUH in 2016. Strictly speaking this script is incomplete, as it simply uses brute force (with no multiprocessing!) and uses no genetic programming tricks. It tries to calculate a 'satisfiability' score for random schedules; even so, the scoring matrices are incomplete. This is put here for historical purposes only.