Initial commit for translate.pl
translate.pl is a Perl script for Wikipedia translators (specifically for English-to-Chinese translations, but easily adaptable for other languages). It prints out the Chinese terms for each intra-wiki link by (very rudely) scraping the contents of each linked page and getting the Chinese page name if it exists. As this method puts extra load on the Wikipedia servers, please do not abuse this script.