Initial commit for solitaire.py
solitaire.py is a Python script that implements Bruce Schneier's Solitaire cipher, as described in Cryptonomicon by Neal Stephenson. It takes a key file (a JSON file containing a list of non-repeating numbers ranging from 1 to 54) and a message file to encrypt or decrypt (file must only contain characters of the alphabet, without whitespace or punctuation). This script was written for fun; do *not* use it for serious encryption.