Added manifest.rb
manifest.rb is a Ruby script for generating manifest files for KRTC card transportation bags. It takes a CSV dump of the produced cards as an argument and creates a folder (on the Desktop) containing manifest files in OpenDocument Spreadsheet (ODS) format for each station. This script relies on a 7-zip executable placed on the Desktop and thus is limited to running on Windows.