Added lotterycheck.awk
lotterycheck is a simple tool for checking if a given set of lottery numbers contains winning numbers. It was written primarily as an exercise in Awk (technically Gawk only, as the asort() function is used here). It takes two CVS files (one for selected numbers and one for winning numbers) and prints the number of matches for each set. First written in July 2013.